runtime/hailo-8/python/super_resolution/requirements.txt
loguru
Pillow
opencv-python